pyproject.toml
[project]
name = "friends-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "google-genai>=1.0.0",
    "ollama>=0.6.1",
    "rich>=14.3.3",
]
